chore(audit): implement all non-cookie audit fixes and update progress tracker
- complete remaining frontend reliability/a11y fixes (WS reconnect unread refresh, modal focus/escape, icon aria labels)
- harden backend auth/cache/logging paths and add missing type hints
- add FK ondelete cascade + bigint PK/FK migrations and align SQLAlchemy models to 2.0 style
- add/adjust tests for new behavior and schema expectations
- update docs/AUDIT.md status (OPEN -> 0; cookie auth remains planned)

+++ backend/app/models/user.py
@@ -1,6 +1,6 @@
 from datetime import UTC, datetime
 
-from sqlalchemy import DateTime, Integer, String
+from sqlalchemy import BigInteger, DateTime, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -10,7 +10,7 @@ class User(Base):
     __tablename__ = "users"
 
     # Columns with proper type hints for mypy
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
     username: Mapped[str] = mapped_column(
         String, unique=True, index=True, nullable=False
     )
